Complete Day 1 and Day 2 Pt 1
This solves Day 1 and Part 1 of Day 2 for the 2019 Advent of Code.

Day 1 has solutions provided for calculating how much fuel your rocket
needs, as well as how much fuel your fuel needs. The part 2 solution
where we calculate fuel for the fuel is implemented using tail
recursion.

Day 2 Part 1 has solutions provided that build a simple Intcode
computer that relies on modifying memory. I chose to implement it with
MutableLists because the program allows for the opcodes of future
instructions to be changed by previous instructions. Therefore, there
was no easy way to do this in a more immutable and functional way.

This also provides all the files to initialize the kotlin implementation
of the advent-2019. This includes all the needed gradle files and git
files.